Allow HTTP S3 Endpoints
I'm currently developing a service which will be using SeaweedFS as an S3 drop in replacement, and I'd prefer to use this library over aiohttp-s3-client as I'm building a FastAPI app, and httpx is already included for the test client. However, the SeaweedFS instance will be running in a kubernetes node and likely won't have SSL. The only way to use an http endpoint atm seem to be modifying `S3Client._aws_client.schema` similar to [one of your tests](https://github.com/samuelcolvin/aioaws/blob/main/tests/test_s3.py#L18), but this isn't exactly ideal. This change allows configuration of `AwsClient.schema`.

If modifying the config isn't desired, it could be changed to use [yarl](https://github.com/aio-libs/yarl) to parse the provided URL and, if a scheme is provided, assign it to `AwsClient.schema`. Off the top of my head:

```
parsed_url = URL(self.host)
if parsed_url.scheme:
    self.schema = parsed_url.scheme
    self.host = str(parsed_url).replace(f'{parsed_url.scheme}://', '')

else:
    self.schema = 'https'
```

I may be missing it, but unfortunately, yarl doesn't seem to have any helper methods to get the URL without the scheme.